Inline IntroPage + css and use box-sizing instead of -webkit/-moz-box-sizing
